Refactor. Move things for OData to Services.Core
