+++ src/ContentRepository/Group.cs
@@ -443,7 +443,7 @@ namespace SenseNet.ContentRepository
         /// <param name="contentIds">An array of contentIds that represents the new members.</param>
         /// <returns></returns>
         [ODataAction]
-        [SnAuthorize(Role = "Administrators, Everyone")] //UNDONE:?? Role = "Administrators, Everyone"
+        [SnAuthorize(Role = "Everyone")]
         [ContentType(N.Group)]
         public static object AddMembers(Content content, int[] contentIds)
         {
@@ -470,7 +470,7 @@ namespace SenseNet.ContentRepository
         /// <param name="contentIds">An array of contentIds that represents the members to remove.</param>
         /// <returns></returns>
         [ODataAction]
-        [SnAuthorize(Role = "Administrators, Everyone")] //UNDONE:?? Role = "Administrators, Everyone"
+        [SnAuthorize(Role = "Everyone")]
         [ContentType(N.Group)]
         public static object RemoveMembers(Content content, int[] contentIds)
         {
