+++ src/OData/Writers/ODataTableWriter.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using SenseNet.OData.Metadata.Model;
+using SenseNet.Services.Core.Operations;
 
 namespace SenseNet.OData.Writers
 {
